im-buddy/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
python-dotenv==1.0.0
requests==2.31.0
supabase==1.0.3
gunicorn==21.2.0
python-decouple==3.8 